Docker, application properties, fixes, docker actions

+++ flight-service/src/main/java/fact/it/flightservice/repository/FlightRepository.java
@@ -3,5 +3,5 @@ package fact.it.flightservice.repository;
 import fact.it.flightservice.model.Flight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FlightRepository extends JpaRepository<Flight, Long>{
+public interface FlightRepository extends JpaRepository<Flight, Integer>{
 }
